feat: csms producer

+++ service/csms/src/main/java/me/jwjung/csms/ocpp/domain/payload/MeterValuePayload.java
@@ -6,7 +6,7 @@ import me.jwjung.csms.ocpp.domain.MessageType;
 @ToString
 public class MeterValuePayload implements CorePayload {
 
-    private Long chargingAmount;
+    private Double chargingAmount;
 
     @Override
     public MessageType getType() {
